Обновление `enviroment` в тестах
