+++ tests/components/the-header/navigation/TheHeaderNavigationCard.spec.ts
@@ -1,3 +1,4 @@
+// @vitest-environment nuxt
 import type { VueWrapper } from "@vue/test-utils"
 
 import { mountSuspended } from "@nuxt/test-utils/runtime"
