+++ tests/constants/personalProjects.ts
@@ -1,0 +1,20 @@
+import { tags } from "./tags"
+
+export const personalProjects = [
+  {
+    description: "Описание личного проекта",
+    github: "https://exer7um.com/",
+    icon: "🍔",
+    id: 1,
+    tags: tags,
+    title: "Личный проект",
+  },
+  {
+    description: "Описание личного проекта",
+    github: "https://exer7um.com/",
+    icon: "🍔",
+    id: 2,
+    tags: tags,
+    title: "Личный проект",
+  },
+] as const
