+++ tests/constants/projects.ts
@@ -1,0 +1,24 @@
+import { tags } from "~/tests/constants/tags"
+
+export const projects = [
+  {
+    backgroundColor: "#000000",
+    description: "Интернет-магазин брендовой одежды",
+    id: 1,
+    logo: "/oonyxx-store.svg",
+    tags,
+    title: "oonyxx.store",
+    url: "https://exer7um.github.io/Oonyxx-Store/",
+    year: 2019,
+  },
+  {
+    backgroundColor: "#000000",
+    description: "Интернет-магазин брендовой одежды",
+    id: 2,
+    logo: "/oonyxx-store.svg",
+    tags,
+    title: "oonyxx.store",
+    url: "https://exer7um.github.io/Oonyxx-Store/",
+    year: 2019,
+  },
+] as const
