+++ tests/server/utils/useLocalizedColumn.spec.ts
@@ -1,3 +1,4 @@
+// @vitest-environment node
 import useLocalizedColumn from "@/server/utils/useLocalizedColumn"
 import { createEvent } from "h3"
 import { IncomingMessage, ServerResponse } from "http"
